Renamed RawPointCloudProvider to StandardPointCloudProvider
